Updated HashMap class
